requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
openai==1.3.0
ollama==0.1.7
pydantic==2.5.0
python-multipart==0.0.6
pystray
pillow
qrcode
python-dotenv
